Aim: To design a system that automatically turns on the lights inside an almirah whenever it is opened and turns off the lights after 100ms if there is enough light in room
link=https://www.tinkercad.com/things/1AePUIt5XOc-magnificent-allis-blorr/editel?tenant=circuits
THEORY:  The concept used for completeing the task assigned was of LDR(light dependent circuit),switch,audrino and breadboard . 
LEARNING AND OBSERVATIONS:  When light falls i.e. when the photons fall on the device, the electrons in the valence band of the semiconductor material are excited to 
the conduction band. These photons in the incident light should have energy greater than the band gap of the semiconductor material to make
the electrons jump from the valence band to the conduction band. Hence when light having enough energy strikes on the device, more and more
electrons are excited to the conduction band which results in large number of charge carriers. The result of this process is more and more 
and hence it is said that the resistance of the device has been decreased.
Problems and troubleshooting:
Loose connection of cables.
Loose connections of LDR device.
Error in programming or coding.
Precautions:
While unplugging the USB ,pull the plug nt the cable.
Connectionns should be tight and according to the coding done on audrino software.
Handle the apparatus like AUDRINO BOARD ,LDR and breadboard carefully.
Learning Outcomes:
Working and appearence of Audrino.
Working of a photoresistor.
Design and analysis of breadboard alongwith its circuit diagram.